rhone: Copying src from kernel tree
All of this code has lived in the private kernel tree repository where
sched_ext is being developed. This commit brings the source code into
the repository here. The code doesn't yet build, and will likely be
moved to a different directory, e.g. src/ and include/ in an upcoming
set of commits.

Signed-off-by: David Vernet <void@manifault.com>